Move addition of double quotes to timer names to GPTL.
The double quoting of timer names is currently implemented only
in perf_mod.F90, and timer names when using the GPTL start/stop
routines directly will not have the quotes added. This commit
removes the logic adding double quotes to timer names in
perf_mod.F90 and adds it as an option for the output of timer
names in gptl.c . Thus internally the timer names do not have
the double quotes. This option is set during the timing library
initialization with perf_mod.F90.

[BFB]