Move detail information to end of timer name
Setting the option profile_add_detail to .true. adds the detail
number associated with a timing event to the timer name (used
for detail debugging, evaluation, and planning). Currently this
is prepended to the timer name within perf_mod.F90. With the
planned move of the timer prefix logic to GPTL, this would place
the detail information between the prefix and the rest of the
timer name. This commit moves the detail number to the end of
the timer name, to make it easier to identify.

[BFB]