Move prefix support from perf_mod.F90 to GPTL
With the inclusion of C/C++ native code in the model
(e.g. PIO2) there will be a need to call GPTL routines directly
and not just via perf_mod.F90 . This commit moves the prefix
support logic from perf_mod.F90 into gptl.c, so that a prefix
set in the model will also modify timer names for timing events
defined using GPTLstart/stop, for example.

This commit also adds support for defining a prefix in a threaded
region, where the prefix applies only to the particular thread.
The current design preserves a prefix defined in a serial region
and appends the prefix defined in a threaded region for that thread.
There are some subtleties in using this, e.g. if a prefix set in
a threaded region for thread 0 is not explicitly unset, it will
continue to be used for thread 0 events outside of the the threaded
region. Thread-specific prefixes is implemented only for OpenMP
threading.

[BFB]