Modificationss made to wsm6 mp physics to unify with wrf wsm6 code
Toward the effort to unify physics schemes between the MPAS and WRF models,
the WSM6 microphysics scheme (module_mp_wsm6.F) was modified to include
changes necessary to allow it to run in both models. In areas where the models
differ (e.g., in their handing of 'dx'), there are if-defs put around the code
to ensure that it's handled correctly for the model in which it is running.
Results before and after are bit-for-bit.