Add Pack and Mask types for vectorization.
At first, these will be used only in a set of unit tests. These unit tests are
meant not just to test correctness at the level of source code, but also to be
used to diagnose compiler problems. Thus, even if some BFB-based regression
tests are not expected to pass in a performance build, the pack unit tests are
expected always to pass, and a failure may indicate a compiler issue.

In this commit, none of the unit tests are implemented. This commit is part of a
precursor branch to bring clean configuration into SCREAM.