Adjust pack_tests for KNL and SKX.
Config: Remove an invalid cmake print_var. Fix ignored CMAKE_CXX_FLAGS.

Add deactivate FPE function. This can be useful in tests that are expected to
have an FPE as a result of vectorization. As it turns out, I'm not using it for
the final set of tests for the following reason:

Be selective in testing; build time with -O3 got long.