Fix incorrect 'xtime' variable for fractional dt
When the model timestep is not an integer number of seconds, the xtime variable
computed in the atm_timestep(...) routine could be short by 1.9999... seconds.

The timestamp string valid at the beginning of the time step is computed from
the exact simulation clock time, and the conversion to a string truncates
the fractional part (i.e., up to 0.99999... seconds). Then, in the atm_timestep
routine, the model timestep, dt, is added to this truncated timestamp.
In the conversion of the resulting time back to a timestamp string, fractional
seconds (i.e., up to 0.99999... seconds) are again lost.

One practical impact of this is that model output and restart files would have
an incorrect 'xtime' field, preventing the model from restarting from the correct
time, e.g.,

 xtime =
  "2010-10-23_23:59:59                                             " ;

To address this problem, we now pass an MPAS_Time_type in place of the timeStamp
character string to the atm_timestep(...) routine. This preserves any fractional
seconds in the beginning time, allowing the computed time at the end of
the dynamics step (i.e., xtime) to be correct when the result has no fractional
part.

Note: There are still other potential problems with fractional seconds.
      For example, if the output interval is not evenly divided by the fractional
      timestep, the 'xtime' field in the output file will contain only
      the integer part of the seconds, and will therefore be up to 0.99999...
      seconds short.