Update config files for LC resources - For maint-1.0 branch
Changes the config files in CIME to reflect recent changes in Livermore Computing resources:
Specifically for the maint-1.0 branch.

Update syrah to a module system to be compatible with a recent OS change
Update quartz to be independent of a spack build
Remove cab as a machine since this machine has been taken down

[BFB]

See confluence for a more detailed description about these tags.