HOMME: Removing traces of the HOMME_FPMODEL cmake variable
We decided to not add this feature for now (it is present in HOMMEXX).
To avoid confusion, I removed all its occurrences from the repo.

Also, the HOMME_GPU_BFB_WITH_CPU compiler definition was added in HOMMEXX
when HOMME_FPMODEL matched 'strict'. Since HOMME_FPMODEL is not present in
here, we turned it into a cmake option (off, by default).