Update RRTMGP coefficients files
Update RRTMGP coefficients files, which have been updated in the
rte+rrtmgp repo, but more importantly these needed to be converted to
netCDF classic format on the data repository instead of netCDF-4 because
some of our machines have trouble reading netCDF-4. This PR updates the
defaults to point to the newly created files on the data repo, which
have already been uploaded. This should fix #2565.