HOMME: abstracting the Context class
The class now only has templated interfaces for accessing stores objects.
That is, no more get_xyz() methods, but rather get<XYZ>().

The reason behind this is extensibility. It is possible that different
models (preqx, theta, ...) require different objects in the Context.
We want to use the same Context object, without having to store
all possible objects for all possible models, even if the current
executable does not need some.

This, however, requires Context to store a map (rather than
individual pointers) that can store 'generic' objects.
The std::any class would provide this functionality, but unfortunately
it is present only in the c++17 standard, which we cannot expect
to be supported on all the target machines. Therefore, we implemented
an in-house 'any' object, in the StdMeta.hpp file. This file is meant
to be 'temporary', implementing things that are in future c++ standars,
or at least foreseen to be there.

The only limitation for objects stored in Context is that they support
a default constructor. No other assumptions are made. This caused a few
adjustments. For instance, the previous implementation offered methods
such as `get_tracers()`, and the impl of that method took care to call
the constructor Tracers(int,int) if the stored pointer had not yet
been initialized. Now the method get<Tracers>() cannot know about the
constructor Tracers(int,int), so the default constructor of Tracers
had to be changed, so to call the more specialized constructor, grabbing
the two int's from other Context classes (Elements and
SimulationParams). Since this could have cause some confusion, we
added a routine called `init_functors_c', which correctly initialize
all the functors in the Context, so that all calls to get<[FunctorName]>()
will simply be 'get' calls, without the need to create a new object.

The Context class can also handle all the stuff that was handled by
MpiContext, de facto removing the need for MpiContext.

The Context class is now *very skinny*, and compiling it does not
force to compile other objects (as it did before). Therefore, there
is no more need for the NO_CONTEXT and NO_MPI_CONTEXT macros, used
in some unit tests before to avoid compiling all the expensive classes
just to access *some* of the stuff stored in Context. Hence, we can
now compile Context even in unit tests. If we only store 1 object
in it, only that object sources will be needed.