Almost scalably initialize regular-mesh cubed sphere.
Currently, HOMME grid initialization is not scalable because the GridVertex and
GridEdge arrays are for the whole mesh. This is the first PR in a sequence of
probably three to address this scalability bottleneck:
  1. Make the regular internally generated cubed-sphere grid initialization
     scalable except for the SFC calculation (this PR).
  2. Make the RRM -- i.e., file-based -- grid initialization scalable.
  3. Make the space-filling curve calculation scalable.
The first is the highest priority. The third is the lowest in priority because
we can have ne > 10 thousand before it becomes a problem, but it will matter for
RRM meshes with very small smallest elements.

Since this PR does 1 but not 3, grid initialization is not completely scalable,
hence "almost".

In this PR, can_scalably_init_grid is determined. If it is true, then the new
scalable implementation is used; otherwise, the old one is used. The new
implementation is in scalable_grid_init_mod. It constructs GridVertex and
GridEdge for owned and remote-used elements only. The subroutine initMetaGraph
cares only about these elements.

Two minor things are also done. Head/TailPartition are not used and so were
removed from prim_init1. Similarly, assign_node_numbers_to_elem is not used and
so was commented out.

Following are results on 32 KNLs for a 3-level model. Number of levels is
orthogonal to grid initialization; I chose 3 so I could fit the memory
associated with the elements on a small number of nodes. Each KNL has 96 GB of
memory. They are run in cache mode. 64 ranks/node are used.

                  Original  Scalable/Phase-1
      ne  grid init1 init1  grid init1 init1
            GB    GB     s    GB    GB     s
     120  0.11  0.11  1.24  0.01  0.01  0.11
     240  0.43  0.43  5.13  0.01  0.02  0.21
     512                    0.01  0.06  0.67
    1024                    0.02  0.20  2.77
    2048                    0.06  0.77 12.23
    2500                    0.09  1.14 17.55
    3000                    0.13

The columns are:
    ne, the cubed-sphere mesh size parameter;
    grid (GB), max resident state size (RSS) in GB after grid initialization;
    init1 (GB), max RSS in GB after prim_init1 has returned;
    init1 (s), the max time over ranks in seconds to run prim_init1.
Entries are blank if the program crashed before the data were available.

The key result is that grid initialization at ne 3000 with the phase-1 code
takes < 0.2 GB/rank, compared with an estimated > 60 GB for the original
implementation. A second result is that prim_init1 is 10-25x faster (based on ne
120 and 240).

[BFB]