Add new temperature-profile-based ice shelf melt param.
This commit adds the basic functionality of an ice shelt basal melt rate parameterization
by Xylar Asay-Davis to be used in the Thwaites variability paper
The parameterization takes a ocean temperature profile and sill depth as input.
The temp. profile is adjusted for cavity temperature behind the sill,
assuming constant temp. below the sill depth.
Then a local temp profile in the boundary layer is estimated assuming exponential
decay from the cavity profile to a plume profile at infinite distance downstream
from the grounding line, accounting for the pressure-melt temperature.
The melt rate is calculated as proportional to the product of the boundary layer
temperature profile and the mean cavity temperature.

Additional functionality is still to be added, including calculating the
average grounding line depth as an input to the parameterization.

At present the parameterization reproduces the behavior of Xylar's
python script.