Update marine b.c. for velocity in Interface
For some unknown reason, one grounded marine point on the Thwaites mesh was being
treated as a terrestrial boundary in the velocity solver interface.  This resulted
in an upper surface elevation well below sea level and associated very steep
surface slope and high velocity.  I was unable to determine why this
location was not being caught by the grounded marine margin check later,
but a simple solution was to initialize the upper surface to the maximum
of the bed elevation + ice thickness and the floatation criterion.