Use mean of 10% of deepest GL depths as GL depth for melt param.
This commit changes the calculation of the GL depth used by the melt
param. to calculate the mean of the deepest 10% of GL depths.  This is
done in an approximate fashion using 10 m wide depth bins to make this
algorithm efficient in parallel.  The depths are also weighted by the
sqrt of cellArea to account for different grid spacing.