Add new mpas_atm_run_compatibility() to verify input fields and namelist options
This commit introduces a new routine in the atm_core module for checking
the consistency and compatibility of inputs supplied at run-time. This routine
is called after packages (and physics suites) have been set up, and also after
namelists and input streams have been read.

In its initial implementation, this subroutine checks that we have what appear
to be valid GWDO static fields if the config_gwdo_scheme options is not 'off'.