Address the SFC part of scalably initializing regular-mesh cubed sphere.
This commit scalably initializes the space filling curve in the case that ne is
factorable. For each owned SFC index range on each face, it finds the start of
the curve, then follows it to the end of the index range. Finding the first
point is a log(ne) operation, and following it to the end of the range is
O(owned elements). O(ne) memory is used instead of the O(ne^2) if the SFC is
created for a whole face.

The main implementation detail is the new sfcmap_* routines in
spacecurve_mod. scalable_grid_init_mod was then modified to use these.

If ne is not factorable, the master process outputs a message giving the next
smaller and next larger ne that are factorable, in case the user wants to switch
to one of these. Then it uses the original method to do the SFC calculations.

At ne=2048 on 32 KNL, prim_init1 takes 9.4s (scalable GridVertex/Edge and SFC)
instead of 12.8 (scalable GridVertex/Edge but not SFC). sgi_init_grid itself
takes 0.43s of the 9.4s. Thus, the speedup in sgi_init_grid is approximately 7x,
but overall this is a small gain. But the main thing is to exercise these
capabilities for use in the SFC step for RRM meshes with very small elements,
the topic of a future PR and part of phase 2.

[BFB]