CIME: reverting homme.py to pre-PR #2531 version
There are some issue with testing. To avoid blocking other PRs,
we revert changes to the testing system to the state it was
before PR #2531, so other PRs can go in while we fix the testing
system.
Note: after this commit, *no testing* will happen in nightlies
for the preqx_kokkos target in HOMME.