Control ouput of MPI task to compute node mapping data
A recent PR fixed the output of component-specific MPI task to compute
node mapping for ATM and LND, and added it for ROF, OCN, ICE, and
GLC. The utility of this output for each component is unclear, and a
namelist variable to enable  or disable it is added here. This
variable also allows setting  the format of the output. All component
model task-to-node mapping output is controlled by this single
cime_pes namelist variable info_taskmap_comp, where setting this to 0
disables output,  setting it to 1 uses the more compact format, and
setting it to 2 uses the more verbose format.

Note that a separate component-only PR will be needed for the
components to be aware of info_taskmap_comp and use this as described
above, but this additional code has been used to test this PR.

For symmetry, a cime_pes namelist variable info_taskmap_model
was also added to control the task-to-node mapping output for
the full coupled model, with the same options (0, 1, and 2). The
default here is 2 (preserving the current behavior).

Also for symmetry, output of task-to-node mapping data was added
for the CPL component (and is controlled by info_taskmap_comp the same
as for the other component models).

Finally, the implementation for the whole-model task-to-node mapping
data output was modified to better reflect the (currently incomplete?)
support for running the model with multiple drivers.

[BFB]
[NML]