Pull in read routine from rte-rrtmgp source
Pull in coefficients read routine from the rte-rrtmgp external source. This seems to fix a problem with
reading the new updated coefficients files. Pulling in those routines is also probably a good thing
because it should help ensure some consistency between the coefficients files provided by rte-rrtmgp
and the code used to read and initialize the gas optics objects.