Pass effectivePressure to Albany velo solver
This will allow the Schoof basal friction law to be used in Albany.  All
the changes are in place, but the actual passing of N is commented out
until the corresponding changes on the Albany side are available.
Selection of the Schoof basal friction law and associated parameter
values will happen in the albany_input.xml file.