Removed the Factory abstract class.
It was not really needed. I replaced it with a tiny struct called
AtmosphereProcessFactory, with a single static method 'create'. Although
it sounds the same as the old Factory class, it is much simpler, pretty much
equivalent to a hand-rolled

  if (name=="bla1") {
    return new Bla1();
  } else if (name=="blah2") {
    return new Bla2();
  }

The reason for isolating the creation of processes in a small class is
explained in the header file control/atmosphere_process_factory.hpp