Upgrading the reinterpret_field function to allow different scalar types.
The routine can now reinterpret f1 into f2 as long as the memory size of
the two data types are the same. This allows to reinterpret, say, a field
with data type 'double[16]' into a field with data type 'Pack<double,4>[4]'.

Note: technically this also allows to reinterpret a field with data type
      'double[4]' into a field with data type `float[8]`, so use with care.