Do not include time of day in cpl auxiliary hist file names
Having the time of day in these file names is incompatible with the file
naming assumption in dlnd. In talking with @mvertens, we felt it best to
back out this change for now (which was made in ESMCI/cime#2782), rather
than changing the convention in dlnd - partly because the old convention
may be assumed elsewhere, too.

Fixes ESMCI/cime#2912