Do not require masks or attributes
Remove expectation that masks are always present, since this concept is irrelevant for newer
unstructured grids. If a mask variable exists, it will be used. Otherwise, it is assumed that the
grid is defined over the entire domain. Also remove dependence on certain global attributes that may
or may not be present.