Adding struct FieldAllocProp, and changin fields registration mechanism
The struct holds some information needed by Field to establish how
much memory to allocate. When fields are registered in the repo,
there is the possiblity of specifying the scalar type that will be
used to access the field. For instance, say we have a 1d field of
9 Real's. One may want to use the field in a vectorized way, with
vector length 4. A request for scalar type Pack<Real,4> for the field
allows Field to figure out that it will need to allocate 12 Real's
rather than just 9, in order to allow correct padding.

To allow this flexibility, we need to pass a scalar type as template
argument of the registration process. So far, the driver was getting
the fields identifiers form all the atm processes, and use them
to register the field. This is no longer enough, and since we cannot
get a data type at runtime, the atm processes have to register the
fields in the repo themselves. Hence, we now pass the repo to all
the atm processes, so that they can register the fields. We also
do some checks to make sure that processes are only registering fields,
and not messing with its internal state (e.g., by calling
registration_complete).