Update system-specific job progress monitoring scripts
When developing the job progress monitoring script for Summit,
some improvements were identified that apply equally to the
scripts for the other systems (syslog.<system-name>).

On Summit, background processes ignore kill signals while they are
asleep, and job scripts will not complete until spawned background
processes are killed or complete normally. Thus, if the job progress
monitoring script sleeps for a long time (a user-defined interval),
allocation time will be wasted. The workaround was to divide the sleep
interval into a loop over 2 minute sleep intervals. This same change
of implementation is implemented here for all systems, to make sure
that background jobs do not outlive the spawning job scripts.

Currently, some system status information is collected twice (with
output of the second call overwriting that of the first). A second
clean-up is to remove this redundancy.

The e3sm log file is copied to the checkpoint directory
at the beginning of the run, in an effort to capture the task-to-node
mapping. However, this copy needs to be delayed long enough for the
mapping information to show up in the log file first. Various methods
have been tried to calculate the number of lines in e3sm.log that would
indicate that this information has been captured. Yet another approach
is implemented here, using empirically determined system-specific
and job-specific line counts. (Note that we cannot finish the update
for Summit until Summit is available for testing again.)

Finally, since Mira and Cetus are no longer target systems for the E3SM
project, the corresponding syslog scripts are deleted.

[BFB]