Update logic on when to copy over e3sm.log file for Summit
Logic on when to copy over the e3sm.log into the checkpoint directory
has been modified for Summit, similar to the modifications for the
other systems. Part of this new logic was then commented out because
we can not yet test this on Summit. When Summit is available for user
codes, probably after January 1, we will test all of the new code
and uncomment it out if it works as expected.

[BFB]