Added a check for FieldAllocProp
When we call 'request_value_type_allocation' we now also check that
the input template argument 'ValueType' is either the same as the
stored scalar type (we don't have the actual type, so we check the
name and the size), or a pack type. This should help ensure that
the only allowed types are ScalarType and Pack<ScalarType,N>.
Theoretically, packs of packs (of packs...) are also allowed, but
there should not be any issue with that, so we don't check that.