Improve error/exception handling throughout CIME
Our default CIME error was SystemExit, which was nice for
avoiding tracebacks, but very bad in that it's not a subclass of
Exception. This forced us to use bare excepts and catches for BaseException.

The new CIME error is CIMEError, which is both a SystemExit error and
an Exception.

There are still a few places where we catch BaseException, these are
places where we want to do cleanup, even in the presense of a KeyboardInterrupt.
These places should almost always re-raise after cleanup because keyboard
interrupts should not be swallowed.

Get rid of all bare excepts.