Mods to build with IAC stub
These are the modifications necessary to build with the iac component stub;
there's a doc file included, and, for some reason, the some mpas submodules
that I have to commit for reasons I don't completely understand.

I have not yet successfully run the developer unit tests on this version of
the code, so there may yet be some bugs that I have to root out, but at
least these changes allow the build to succeed, and contain the bulk of the
modifications and refactorings necessary to add an eighth component.

(Also corrected my name and email from the git config)

*********1*********2*********3*********4*********5*********6*********7**
Longer commit message body describing the commit.
Can contain lists as follows:
	* Item 1
	* Item 2
	* Item 3

A good commit message should be written like an email, a subject
followed by a blank line, followed by a more descriptive body.

Can also contain a tag at the bottom describing what type of commit this is.
[BFB] - Bit-For-Bit
[FCC] - Flag Climate Changing
[Non-BFB] - Non Bit-For-Bit
[CC] - Climate Changing
[NML] - Namelist Changing

See confluence for a more detailed description about these tags.