Restore how omega is output to history file for EUL-SCM
PR # 2667 introduced a change to have a more consistent treatment of calling outfld to save OMEGA field works for SE dycore. But after the change, the prescribed vertical velocity field would not be saved as is because of a currently unidentified issue in EUL-SCM concerning state%omega. This causes integration test SMS_R_Ld5.T42_T42.FSCM5A9 to have DIFF failure, though the simulation results are not affected. The original format of outfld call is restored until a fix for EUL-SCM on state%omega is found.

[BFB]