Allow setting fminval and fmaxval on command line
Allow setting fminval and fmaxval through the command line to specify the
minimum and maximum allowable land fractions. Land fraction below fminval is
reset to 0, and land fraction above fmaxval is reset to 1. These options may be
necessary when dealing with coarse grids. For example, in computing domain files
from mapping files between the oQU240 and ne4 grids generated using
TempestRemap, setting fminval > 0.001 was necessary to get a land domain mask
that looked similar to those generated from ESMF mapping files. Setting fmaxval
< 1 was also required to get ocean domain masks that looked reasonable.
Regardless, this should be a user-adjustable setting, and adding these as
command line arguments accomplishes this. The defaults of 0.001 and 1.000 are
maintained if the arguments are omitted.