Added history fields for atmospheric flux variables to v2 data structs
Completed the addition of new history variables for the v2 data
architecture at the topounit level. All previous history field names
and other metadata are maintained in the new code, so this change is
BFB with master. At a later date, ELM team may want to add additional
history outputs, or remove some that are duplicates with different names
as part of the code cleanup process.

[BFB]