Error-check the cprnc output
Fail if we can't find one of the expected strings in the cprnc
output. Point is: if the cprnc Fortran code is changed to change the
format of one of these strings without corresponding changes in the
python, we want to fail rather than silently doing the wrong thing.

Also refactor this logic a bit.