Homme: moved CaarFunctorImpl.hpp to preqx_kokkos/cxx
It is a preqx-specific implementation. On the other hand, CaarFunctor.*pp
can stay in share/cxx, since it does not contain model-specific stuff.
Although in F90, for the theta model, compute_and_apply_rhs is actually
called compute_andor_apply_rhs, for the sake of minimizing code duplication,
we keep the base class in share/cxx, and move only the implmentation one.