Allow specifying directives to use for an unknown queue
Previously, when using an unknown queue, the batch directives would come
from the default queue for the machine. This is a problem on cheyenne,
for which the default queue is the share queue. It looks like we need to
keep the share queue as the default queue for other purposes, though (in
particular: if the default is the regular queue, then the share queue
won't get selected for small - e.g., single-processor - jobs, since
small jobs also satisfy the requirements of the regular queue), so the
solution is to provide a different mechanism for specifying the queue to
use for the sake of choosing directives.

For machines that don't use this new mechanism, we maintain the previous
behavior of using the directives for the machine's default queue.

Fixes ESMCI/cime#2994