Homme: Split Diagnostics into base and derived classes.
The base classes only implements prim_diag_scalars, which is the same in
both preqx and theta. The derived one takes care of implementing prim_energy_halftimes,
which differ in the two models. The new design requires the freshly added
feature of Context, that allows to register and access to a product with both
its type name and one of its base types' name.

# Conflicts:
#	components/homme/src/preqx_kokkos/CMakeLists.txt