Require command line flag to infer masks
Require a command line flag (set-omask) to enable inferring ocean mask
when it is not present in the input mapping file. By default, if mask_a
is not present in the input mapping file, the code will throw and error
and exit. However, if --set-omask is passed, then in the absence of
mask_a, it will be inferred by setting the mask to a vector of 1s, which
are then remapped onto the land grid. This option is only appropriate
for unstructured grids which cover only the area in which the ocean is
active. I.e., this is NOT an appropriate thing to do for regular lat-lon
grids, but should be fine for MPAS-O grids. Tested with a mapping file
from oRRS30to10 to ne30np4 and confirmed these two modes of operation
(using provided mask, and omitting the mask and passing --set-omask to
calculate it) give the same answers.