Do not assume relative path for nf-config
Do not assume a path for `nf-config` relative to `LIB_NETCDF`. This was causing a
problem because different machines would have a different location for `nf-config`
relative to `LIB_NETCDF`. So calling `nf-config` without specifying a path (and
assuming that `nf-config` exists in the `PATH` environment variable) seems to make
the build more portable and not require hand editing the Makefile.