Fixes indexing issue in SurfaceAreaWeighted AM
Nothing is presently being stored in workArray(1,:) leading to all
values in the var_array being shifted by one index.  This is causing
MPAS-Analysis failures. This results in avgSurfaceTemperature values
being stored in avgSurfaceSalinity in output files.

This commit shifts indices back one value so the ordering is as
expected.