Refactored a bit the dynamics folder, to accommodate the use of homme
When using homme, dynamics can be built as a library only if the
configure-time parameters for homme are known (NP, PLEV, QSIZE,...).
This forces us to build dynamics only just-in-time, for each executable.
We can also contemplate the idea of pre-compiling some configurations, which
could reduce compilation time if several tests use the same homme configuration.
This could be a follow-up work to this.

Note: we add homme as a subdirectory, so we need to pre-define some cmake
      variables that homme standalone handles, but that homme as a subdir
      doesn't. We also need to manually handle some work that the homme
      cmake macro createTestExec was doing (e.g., configuring the config.h
      file, or linking tpls, or setting the modules folder).