Modified scream system tests structure.
Now that dynamics is built just-in-time, we need to build it based on the
particular test configuration for homme (assuming homme is the dycore for
the dynamics). I set up a folder with an example of how to handle the cmake
logic, even if dynamics is not really used in the test stub.