Working version of pg1 & pg2!!!
This commit includes the major changes that were needed to get a working version of the 1x1 (pg1) and 2x2 (pg2) FV physics grid. This included a major overhaul of dyn_grid.F90. There appears to still be a bug in the code since the output data has an odd checkerboard pattern. Also, these changes will continue to be updated because we would prefer to encapsulate the mapping between FV physics and SE dynamics grids intoa  new module, which could then hold alternate mapping techniques.

	modified:   components/cam/src/dynamics/se/dp_coupling.F90
	modified:   components/cam/src/dynamics/se/dyn_grid.F90