Avoid unpredictable values in metric terms zb, zb3 for boundary edges
This commit adds logic to set the index of non-existent cells -- those on
the exterior side of a boundary edge in a limited-area mesh -- to the same
as the cell that does exist. This is done only to avoid bad values in
the computed z-metric terms zb and zb3 for boundary edges.

The exact value assumed in the computation of zb and zb3 for edges on
the outermost boundary of a limited-area mesh should not matter, since
these edges are specified-zone edges that border specified-zone cells.