Removed write_grid_scrip() method
This routine was created as a simple way to get scrip files for pg2 grids, but we can now easily generate these with TempestRemap. At first these scrip files were slightly different from the grid that HOMME generated, but as long as the atm namelist includes "cubed_sphere_map = 2" the grid coordinates are the same.

	deleted:    components/cam/src/control/write_grid_scrip.F90
	modified:   components/cam/src/dynamics/se/inital.F90