Apply maximum MPI tag value patch
In ESMCI/cime#2804 and ESMCI/cime#2892, the MPI tag generation for MCT
was adjusted to allow more than ~30 multi-instance runs fixing
E3SM-Project/E3SM#2487.

This patch is required in order to run Salil's climate reproduceability
tests on maint-1.0, to determine the reporoduceablility of simulations
on Cori which were originally run on Edision.