Further cleans up the post-processing done after p3_main is called in the p3_interface routine.
Unnecessary calculations were removed.
Unnused variables removed.
Calculations of pbuf variables that are passed to other atmospheric
processes have been verified.
A major change was the removal of all variables associated with
	CARMA
	SILHS
	and heterogenous freezing.
These were removed because we don't intend to support them in SCREAM, and
if they are added later that action will be done with much more attention
to detail ensuring that their use is consistent with P3.

This commit also changes the configuration files for the LC machines
syrah and quartz.  A recent change to the default modules on these
machines stopped the code from compiling.