Update config files for LC resources
Changes the config files in CIME to reflect recent changes in Livermore
Computing resources:

Update syrah and quartz machine config entries to explicitly load
mvapich2/2.2 which is the only mvapich version on LC that grants
access to pnetcdf/1.9.0.

These changes reflect a change in the default module loads on LC
machines.

In addition, there is a change to the run_script to purge all mention
of the machines cab and sierra (which have been taken down) and to
add a warning message to the build section warning of the sensitivity
of CIME to double slashes (//) in the case directory path.

[BFB] - Bit-For-Bit

See confluence for a more detailed description about these tags.