A class that manages the grids.
This class serves the purpose of hiding the implementation of how we create
and initialize the grids from the AtmosphereDriver (AD). This means that
the driver need not knowing how the grids manager is created, nor how it
creates/initializes the grids.

The grid manager is responsible for providing a grid object (actually, a
shared pointer) for each of thr grid types.

We provide a simple implementation, called UserProvidedGridsManager, which
does not build grids, but instead relies on the different grids being
set in it. This is used in simple unit tests, and allows to avoid dealing
with the complex initialization of grids, which will be needed in real
simulations.