Change in the internal storage of the FieldRepository
The repo now stores a map of maps. The outer map has as a key the bare
name of the field (e.g., "temperature"). The inner map uses the field
identifier as a key. This means that two fields with the same name but
different layouts (or grids, or dimensions) will be stored in the same
entry of the outer map, but in different entries of the inner map.

This distinction will be used by the Atmosphere Driver (AD) to build
remapping objects to remap corresponding fields with different layouts.