Change default MODIS albedo supersampling factor from 1 to 3
The supersampling factor previously used for the MODIS 0.05-degree monthly
albedo and maximum snow albedo was 1. For finer MPAS-A meshes, e.g., those
refining down to 3-4 km resolution, a larger supersampling factor must be
used.

In order to reduce the number of namelist changes that users must make when
producing static fields for refined MPAS-A meshes, the default supersampling
factor is now 3, which gives about 1.85-km pixels along the equator and has
been found to be adequate for 3-km refinement.

The supersampling factor option now appears in the default namelist.init_atmosphere
as well.