Bug fix in homme's phys-dyn remapper
The dimension of vector/tensor fields must be taken from the field layout
(from inside the field identifier), and *not* from the reshaped view.
In fact, the reshaped view may contain padding (in case the field had
to accommodate pack-type scalar types), and the padding is likely to be
different in the two fields (physics and dynamics). For instance, a
2d vector field with 2 components, with pack<real,4> scalar type support,
will have no padding in the dynamics layout (the last dim is NP=4,
so no padding needed), but *will* have padding in the physics layout
(where the last dim is 2, so there will be 2 double's worth of padding)