Adjust CPL restart i/o routines to handle new changes
The OCN-GLC time averaging accumulators should only be read/written to
coupler restart files if OCN-GLC are actually coupled.  Otherwise there
will be an attempt to read/write unallocated things.

This required passing the ocn_c2_glc flag to the i/o routines, because
that information is not in infodata.