Add's rtype to all constants used in P3 to ensure consistency in precision calculations.
This commit is non-BFB in that prior code did not enforce precision everywhere
so some of the values may have changed.  B4B consistentcy was checked by forcing
a baseline version of the model to be fully double precision using the intel
fortran compiler option '-fpconstant', then making the changes and comparing.

Thus, when all constants are set to be consistent in terms of precision this commit
is BFB.

Finally, there was a handful of constants that were non-BFB even when precision was
enforced by compiler flag.  These haven't been changed in this commit and will be
addressed in a later commit.  These are clearly marked by comments using the line
"rtype problem".