Add lbcs package, lbc stream, and lbc_state pool for LBC init case
This commit defines an 'lbc' output stream for writing model LBC files,
an 'lbcs' package to control the writing of the 'lbc' stream and the allocation
of the 'lbc_state' variables, and a new var_struct defining the fields
that are needed in MPAS-A LBC files.

Code has been added to properly set up packages when init case 9 is run,
and appropriate calls to the MPAS stream mananger for the 'lbc' stream
have also been added.