Generalize logic so that handling of boundary scalars works for any physics suite
The scalars are reset along the domain boundaries individually (i.e., for qv, qc,
etc.) after scalar transport and after microphysics. Previously, there was no
logic to detect which scalars where actually active and which were not, so it was
necessary to comment or un-comment, e.g., the handling of nr and ni, depending
on which physics suite was being run.

Now, there is a check to see that the index of each scalar is valid before setting
that scalar along the domain boundaries.

Also, it appears that the last argument to mpas_atm_get_bdy_state in the previously
commented-out calls to mpas_atm_get_bdy_state for nr and ni had an incorrect final
argument in the call to mpas_atm_get_bdy_state: this argument should be 'dt' rather
than 'rk_timestep(rk_step)'.