Avoid harmless warning messages in mpas_atm_get_bdy_state for scalar constituents
When the mpas_atm_get_bdy_state routine is called for scalar constituents
(e.g., qv), the calls to mpas_pool_get_array will not find the LBC field in
the tend and state pools as a standard field, leading to warning messages when
the code is compiled in debug mode. These messages are expected and harmless,
so this commit adds calls to set the error mode for the pool routines to
MPAS_POOL_SILENT only for these calls to mpas_pool_get_array.