Add subroutine call to get PIO format
Add call to `shr_pio_getioformat` to get PIO format in the event that `mode_in`
is not passed to `cam_pio_createfile`. This should be the default use, and
callers should probably *not* pass `mode_in` to override this in most cases.
However, we leave the `mode_in` argument in case callers need to override the
default type for whatever reason.