Add optional argument 'unitNumbers' to core % setup_log(...) routine
The mpas_log_init routine takes an optional argument, unitNumbers, to specify
which Fortran unit numbers to use for the output and error log files written by
a core. The mpas_log_init routine is called by the core % setup_log routine,
which previously did not have an argument for passing in logging unit numbers.

This commit adds an optional argument, unitNumbers, to the core % setup_log
routine, and it modifies the implementation of each core's setup_log routine
to pass the optional unitNumbers argument to the mpas_log_init routine.

The core % setup_log routine is called by the MPAS subdriver's mpas_init
routine, and at present no changes are made to the mpas_init routine to pass
optional unitNumbers to the core % setup_log routine. However, the changes in
this commit do enable alternate implementations of mpas_init to supply unit
numbers to be used for logging, e.g., in coupled systems where the coupler
has already defined logging units for each component.