Recompute initial RH w.r.t. ice below freezing for consistency with model output
The 'relhum' field in the model initial conditions file was directly
interpolated from the input intermediate file, which contains relative humidity
with respect to liquid water as modified by the WPS ungrib program.
However, the 'relhum' field in the model output is computed with respect to ice
for temperatures less than 273.15 K.

This commit modifies the 'relhum' field in the model initial conditions
to be with respect to ice below 253.15 K, and a transition to RH w.r.t.
liquid water above 273.15 K.

These changes should enable the model initial conditions files to be more useful
for comparison against model forecast fields.