Verify that the input_interval for the 'lbc_in' stream is not 'none'
The mpas_atm_bdy_checks routine now checks that the input interval for
the 'lbc_in' stream can be retrieved and that it is not 'none'. It is
assumed that if the interval exists and is not 'none', then it represents
a valid time interval. Invalid time interval strings currently trigger
a crash when setting up streams.