Check for a valid vertical grid in init_case 9 (LBC setup case)
To avoid trying to produce LBCs for limited-area MPAS-A without a valid
vertical grid, add a check in init_atm_case_lbc on the maximum value of
the zgrid field. If the global maximum of zgrid is zero, it is very likely
that the input stream did not contain valid vertical grid information.