Major update and bug fix
There are a lot of changes in this commit. The ability to output and restart from "cam.i" files was fixed. There were also a lot of cosmetic updates to the code and comments.

There was also a bug in how the tracer tendencies were communicated form physics to dynamics, which is fixed by this commit. The way I implented the fix required adding an initial q state variable (qo) to the physics state structure. This new variable stores the initial tracer state interpolated from dynamics so that it can be subtracted at the end of the physics to get a tendency. The dynamics q state is then added in after interpolating to the GLL grid so that the constituent tendency is consistent with the state of each GLL node.

	modified:   components/cam/src/dynamics/se/dp_coupling.F90
	modified:   components/cam/src/dynamics/se/fv_physics_coupling_mod.F90
	modified:   components/cam/src/dynamics/se/stepon.F90
	modified:   components/cam/src/physics/cam/physics_types.F90